% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genome.R
\name{avgProteinPerCh}
\alias{avgProteinPerCh}
\alias{avgProteinPerCh,Genome-method}
\title{Average Proteins Per Chromosome}
\usage{
avgProteinPerCh(object)

\S4method{avgProteinPerCh}{Genome}(object)
}
\arguments{
\item{object}{Genome object}
}
\value{
Rounded average proteins per chromosome
}
\description{
Calculates average number of proteins per chromosome
}
